Refactor to reduce the setup time for all but one ItopDataTestCase
